new/oss-fetcher/requirements.txt
requests==2.32.3
influxdb==5.3.2
python-dotenv==1.1.1